src/prediction/requirements.txt
fastapi
fastapi-health
uvicorn
pydantic[dotenv]
sentry_sdk
newrelic
openai
langchain
pandas
python-dotenv